+++ cogs/ChallengeReactionCog.py
@@ -1,17 +1,15 @@
-from datetime import datetime
 from urllib.parse import urlparse, parse_qs
 import discord
 from discord.ext import commands
 from Room import Room
 from Request import Request
 
-from const_messages import CHALL_ME_TITLE
+from const_messages import CHALL_ME_TITLE, MAX_GAMES
 
 
 class ChallengeReaction(commands.Cog):
     def __init__(self, client):
         self.client = client
-        self.ehhh = False
 
     @staticmethod
     def is_it_this_request(req, msg, user):
@@ -19,7 +17,7 @@ class ChallengeReaction(commands.Cog):
             return False
         if msg.id != req['req'].message.id:
             return False
-        if user.id != req['req'].defender.id:
+        if user.id != req['req'].defender.id and user.id != req['req'].attacker.id:
             return False
         return True
 
@@ -33,6 +31,14 @@ class ChallengeReaction(commands.Cog):
                 return True
             
         return True
+
+    async def max_challenge(self, author, user, reaction):
+        count = await self.client.usefulCogs['DB'].game_count(author.id, user.id)
+        if count >= 5:
+            await reaction.remove(user)
+            await user.send(MAX_GAMES)
+            return True
+        return False
 
     def get_embed_author(self, message):
         if message.embeds[0].author.name != "Mystery Challenger":
@@ -49,17 +55,16 @@ class ChallengeReaction(commands.Cog):
             if self.is_it_this_request(req, react.message, user):
                 return req
         ## Now checking if it is a 'challenge me' message
-        if react.emoji != self.client.usefulBasicEmotes['yes']:
+        if react.emoji != self.client.usefulBasicEmotes['yes'] and react.emoji != self.client.usefulBasicEmotes['no']:
             return None
-        self.ehhh = True
+
         if react.message.embeds and react.message.embeds[0].title == CHALL_ME_TITLE:
             author = self.get_embed_author(react.message)
-            if author.id == user.id:
-                return
             if not author:
                 return None
             req = Request(react.message, author, user)
-            return {'req': req, 'time': datetime.now()}
+
+            return {'req': req, 'chall_me':True}
         return None
 
     async def create_room(self, msg, attacker, defender):
@@ -74,16 +79,21 @@ class ChallengeReaction(commands.Cog):
         request = self.get_valid_request(react, user)
         if not request:
             return
+        author = self.get_embed_author(react.message)
         if react.emoji == self.client.usefulBasicEmotes['yes']:
-            author = self.get_embed_author(react.message)
             if author.id == user.id:
+                return
+            if 'chall_me' in request and await self.max_challenge(author, user, react):
                 return
             await self.create_room(request['req'].message, request['req'].attacker, request['req'].defender)
             if request in self.client.DuelRequests:
                 self.client.DuelRequests.remove(request)
         elif react.emoji == self.client.usefulBasicEmotes['no']:
+            if 'chall_me' in request and author.id != user.id:
+                return
             await request['req'].message.delete()
-            self.client.DuelRequests.remove(request)
+            if not 'chall_me' in request:
+                self.client.DuelRequests.remove(request)
 
 def setup(client):
     client.add_cog(ChallengeReaction(client))
